requirements.txt
tqdm==4.41.1
musdb==0.4.0
torchaudio==0.7.2
librosa==0.8.0
